build: add basic meson support

+++ src/lxc/string_utils.h
@@ -3,23 +3,23 @@
 #ifndef __LXC_STRING_UTILS_H
 #define __LXC_STRING_UTILS_H
 
-#include <stdarg.h>
-
 #include "config.h"
+
+#include <stdarg.h>
 
 #include "initutils.h"
 #include "macro.h"
 
 #ifndef HAVE_STRLCAT
-#include "include/strlcat.h"
+#include "strlcat.h"
 #endif
 
 #ifndef HAVE_STRLCPY
-#include "include/strlcpy.h"
+#include "strlcpy.h"
 #endif
 
 #ifndef HAVE_STRCHRNUL
-#include "include/strchrnul.h"
+#include "strchrnul.h"
 #endif
 
 /* convert variadic argument lists to arrays (for execl type argument lists) */
@@ -151,13 +151,13 @@ static inline char *deabs(char *str)
 	return str + strspn(str, "/");
 }
 
-#define strnprintf(buf, buf_size, ...)                                            \
-	({                                                                        \
-		int __ret_strnprintf;                                             \
-		__ret_strnprintf = snprintf(buf, buf_size, ##__VA_ARGS__);        \
-		if (__ret_strnprintf < 0 || (size_t)__ret_strnprintf >= buf_size) \
-			__ret_strnprintf = ret_errno(EIO);                        \
-		__ret_strnprintf;                                                 \
+#define strnprintf(buf, buf_size, ...)                                                    \
+	({                                                                                \
+		int __ret_strnprintf;                                                     \
+		__ret_strnprintf = snprintf(buf, buf_size, ##__VA_ARGS__);                \
+		if (__ret_strnprintf < 0 || (size_t)__ret_strnprintf >= (size_t)buf_size) \
+			__ret_strnprintf = ret_errno(EIO);				  \
+		__ret_strnprintf;                                                         \
 	})
 
 static inline const char *proc_self_fd(int fd)
